WIP: Add GitHub remote url parser
Make GitHubUrlParser a non-static class, split class into multiple files, add missing documentation comments
